backend/requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
pandas==2.2.1
python-dotenv==1.0.1
Flask-CORS==4.0.0 